The meaning is different as verb and noun
The word harvest in Russian as a noun is different than verb. 

The old translation "Ожидание урожай" means like we wait for crop to grow and ripe. There is not harvest ready to collect but we are waiting for harvest to be ready. 
My version say that there is already a harvest ready and we are waiting not for a harvest itself but for the action to collect it.