Not correct meaning
Here, translation was you can exchange back. So I think if stake is translated as вложение then unstake изъять вложение. This will be more clear especially for new uses who is not get used to all this financial terms. Even for and users it is not consistent when you stake it in but exchange back out.