feat: add exports for motion components (#10393)
<!--
Before opening a pull request, please read the [contributing
guidelines](https://github.com/pancakeswap/pancake-frontend/blob/develop/CONTRIBUTING.md)
first
-->

<!-- start pr-codex -->

---

## PR-Codex overview
The focus of this PR is to optimize motion components and exports in the
UI kit.

### Detailed summary
- Added `LazyAnimatePresence` component for optimized framer animations
- Updated exports in `Box` and `Motion` components
- Added `MotionFlex` and `MotionGrid` components for animated flex and
grid styling

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->

## Why Add These Components
there is another PR where i want to add animations that rely on the
`AnimatePresence` component. We already have framer installed in the UI
kit, but it is only used internally in the uiKit for effects on other
prebuilt components.

Instead of needlessly re-installing installing framer in apps/web, i
decided just to add exports for the AnimatePresence motion compoent from
our UI kit sonce framer is already installed here.

## Why Export Animate Presence
There are very limited things you can do with framer without wrapping
animated components inside the `AnimatePresence` wrapper. this wrapper
lets animations animate out when they are removed from the dom. Many
animations rely on it, for advanced usage and its not exported from our
UiKit. Also sinde we export `MotionBox` as an optimized `m.div` instead
of a `motion.div` we dont have the extra features like animations,
layouts etc. so exporting `AnimatePresence` lets us resolve this in
scenarios advanced anims are needed

## Addition of other Motion Components
i also added the `Flex` and `Gird` equivilents for their animated
version via `MotionFlex` and `MotionGrid`. since only exporting
`MotionBox` limits what styles/CSS properties we can enforce on the
motion comp.

## But we Already have the MotionBox Component Why Add These
Yes we already have the Motion box component. but its exported an styled
Box and doesnt accept props for flexdirection etc or alignItems and
other flexbox atyles unless u extend it as a styled MotionBox.

## We Already Have Swiper. Why Do We Need This
Correct. We can already can make animated components with swiper.
However there are some siutations where swiper is not as useful or as
effective as framer-motion, or more tedious to use. For example swiper
sometimes requires the usage of additonal custom hooks to manage event
handlers like `onClick` to trigger certain animations. these hooks have
to be implemented in these scenarios which requires extra code and the
`initiation` of the `SwiperClass`. There are many `framer animations`
that can achieve the same effects without the need for defining such
logic thus allowing for solutions that need much less code.

Whats more is Swiper does not support the animation of elements that are
removed from the dom. for example the exit animation of an element out
of view as the enter animation of another element is coming into view.
So for scenarios where you want to fade out and element at the same time
another one fades in swiper cannot be used. So its best to have the
option to use swiper or framer in the situations they're best used for.

## Summary
These changes of fixing the `motion` import to support Animations and
adding `extra motion elements` aswell as exporting `AnimatePresence`
will make it much easier in future to do complex and elegant animations
throughout or web app.